Refactor find
